major updates to main.py
The update includes the renaming of most variables, implementation of frame smoothing for all actors, program information is now present in the top-left corner of the application, and a proper window title has been added.  Overall, the code has been cleaned up to make it more readable and reduce clutter.